refactor(catalog): route provenance imports directly through provenance.py
The provenance API (CatalogProvenance, CatalogCreationMethod,
CloneDetails, get_catalog_provenance, set_catalog_provenance)
lives in catalog/provenance.py. Internal call sites in
src/deriva_ml/__init__.py (5 lazy-load branches) and
core/base.py (2 imports) were routing through catalog/clone.py
which re-exported them with # noqa: F401.

This was a back-compat shim that the back-compat shim itself
didn't need — the importers are all internal. Change them to
import from catalog.provenance directly and drop the re-export
block in catalog/clone.py.

Audit Phase 3 catalog/ §1.8. Risk: trivial — path swap with
no behavior change. The top-level lazy imports (deriva_ml's
__getattr__ surface) still work identically.
